Better env management & login button polishing
Only the environments that belong to the account the user is connected to
are showing in the environment switcher. The environment switcher is hidden
when the user is not logged in. This makes it possible to enable the
storefront for several Scalr accounts. The environment names are fetched
from Scalr.